atualizacao de regex, api
