+++ projeto-labefood/src/pages/createaccount/styled.js
@@ -136,146 +136,18 @@ border: 0;
 img{
 width: 24px;
 height: 24px;
-margin: -6px;
-margin-top: 8px;
+margin: 25px;
+margin-top: 15px;
 }
 &:hover{
     cursor: pointer;
 }
 `
 export const Header = styled.div`
-width: 360px;
-height: 30px;
+width: 370px;
+height: 60px;
 margin: 0 0 0px;
 padding: 0 0 10px;
 box-shadow: 0 0.5px 0 0 rgba(0, 0, 0, 0.25);
 background-color: #fff;
 `
-/*
-
-
-
-
-export const ContainerPrinc = styled.div`
-   width: 100vw;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-`
-
-export const Container = styled.div`
-  width: 360px;
-  height: 640px;
-  padding: 0 0 0px;
-`
-
-export const InputCamp = styled.div`
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    div{
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        flex-direction: column;
-        input{
-            width: 328px;
-            height: 56px;
-            margin: 10px;
-            font-size: 15px;
-            border: 1px solid grey;
-            border-radius: 5px;
-            background: none;
-            z-index: 0;
-            :focus{
-                outline: 0;
-            }
-            ::placeholder{
-                opacity: 0.4;
-                font-size: 17px;
-                position: relative;
-                left: 16px;
-            }
-        }
-        label{
-            font-size: 15px;
-            position: relative;
-            right: 120px;
-            top: 22px;
-            width: 61px;
-            z-index: 1;
-            background: white;
-            padding-left: 5px;
-        }
-    }    
-`
-
-
-export const Title = styled.div`
-width: 360px;
-height: 42px;
-margin: 16px 0 0;
-padding: 12px 32px;
-`
-
-export const Text = styled.h1`
- display: flex;
- justify-content: center;
- align-items: center;
- width: 296px;
- height: 18px;
- font-family: Roboto solid ;
- font-size: 16px;
-
-`
-
-export const Button = styled.div`
-width: 296px;
-height: 18px;
-font-family: Roboto;
-font-size: 16px;
-letter-spacing: -0.39px;
-text-align: center;
-color: #000;
-transition: 0.5s;
-&:hover{
-    transform: scale(1.05);
-    cursor: pointer;
-}
-button{
-  width: 328px;
-  height: 42px;
-  padding: 12px 16px;
-  border-radius: 2px;
-  background-color: #e86e5a
-}
-`
-
-export const Subtitle = styled.span`
-width: 5px;
-height: 8px;
-margin: 3px 234px 43px 16px;
-font-family: Roboto;
-font-size: 12px;
-letter-spacing: -0.29px;
-color: #b8b8b8;
-`
-
-export const ImageLogo = styled.img`
-width: 150px;
-margin-left: 100px;
-`
-
-export const Input = styled.input`
-width: 300px;
-height: 40px;
-padding: 10px;
-margin: 4px;
-font-family: Roboto;
-font-size: 16px;
-letter-spacing: -0.39px;
-color: grey
-`
-*/
